Make date-picker controlled
Right now this library only supports Hebrew dates - I want to give my users an option of choosing a date via a gregorian calendar as well. To keep it simple I am using two date pickers side by side and want the selection on one to update the other.
The issue is that your value prop is really a defaultValue prop, and doesnt allow controlling the value from the outside. Hence this change to make the calendar a controlled component.
Thanks for sharing such a useful component.